Unstuck footer b/c of scroll() for header.
The HTML was set to height: 100%, and this is used by the jQuery
function scroll() to determine placement within the body (document).
The header could not transition up and down based on the scroll without
unsticking the footer (which requires the html and body to be 100%
height). C’est domage.